Fixed selected polygon coloring when rendering categorical feature
