+++ packages/core/src/tile/model/geometry/geometry.ts
@@ -68,7 +68,7 @@ export class GeometryTile extends Tile<GeometryResponse> {
 
     this.mesh.material = material;
     this.mesh.scaling.z = -1;
-    this.mesh.layerMask = 1;
+    this.mesh.layerMask = 0b1;
     this.mesh.renderingGroupId = 1;
 
     renderTarget.renderList?.push(this.mesh);
